addedd swaagger and complleted docs

+++ src/tasks/tasks.service.ts
@@ -5,6 +5,7 @@ import { Task } from './task.entity';
 import { CreateTaskDto } from './dto/create-task.dto';
 import { TaskStatus } from './task-status.enum';
 import { GetTaskFilterDto } from './dto/get-tasks-filter.dto';
+import { User } from '../auth/user.entity';
 
 @Injectable()
 export class TasksService {
@@ -13,16 +14,14 @@ export class TasksService {
         private taskRepository: TaskRepository,
     ) {}
 
-    getTasks(filterDto: GetTaskFilterDto): Promise<Task[]> {
-        return this.taskRepository.getTask(filterDto);
+    getTasks(
+        filterDto: GetTaskFilterDto,
+        user: User): Promise<Task[]> {
+        return this.taskRepository.getTask(filterDto, user);
     }
 
-    // getAllTasks(): Task[] {
-    //     return this.tasks;
-    // }
-
-    async getTaskById(id: number): Promise<Task> {
-        const found = await this.taskRepository.findOne(id);
+    async getTaskById(id: number, user: User): Promise<Task> {
+        const found = await this.taskRepository.findOne({ where : { id, userId: user.id}});
         if (!found) {
             throw new NotFoundException(`Task with ID "${id}" not found`);
             }
@@ -30,12 +29,15 @@ export class TasksService {
 
     }
 
-    async createTask(createTaskDto: CreateTaskDto): Promise<Task> {
-       return this.taskRepository.createTask(createTaskDto);
+    async createTask(
+        createTaskDto: CreateTaskDto,
+        user: User,
+        ): Promise<Task> {
+       return this.taskRepository.createTask(createTaskDto, user);
     }
 
-    async deleteTaskById(id: number): Promise<string> {
-       const result = await this.taskRepository.delete(id);
+    async deleteTaskById(id: number, user: User): Promise<string> {
+       const result = await this.taskRepository.delete({ id, userId: user.id});
 
        if (result.affected === 0) {
         throw new NotFoundException(`Task with ID "${id}" not found`);
@@ -43,8 +45,12 @@ export class TasksService {
        return `Deleted ${result.affected} entries`;
     }
 
-    async updateTaskStatus(id: number, status: TaskStatus): Promise<Task> {
-        const task = await this.getTaskById(id);
+    async updateTaskStatus(
+        id: number,
+        status: TaskStatus,
+        user: User,
+        ): Promise<Task> {
+        const task = await this.getTaskById(id, user);
         task.status = status;
         await task.save();
         return task;
